'quantizedDownsample' option for smoother scrolling
When downsampling, the indices chosen to bucket over change
whenever the x data is updated.  This causes the plot line
to 'wiggle' when plotting a real-time data stream.  The
quanitized downsampling option snaps the bucket centers
to deterministic points such that the downsampled values
for the same underlying data don't change as new data is
added.  This updated also switches to using midpoint x-values
for downsampled data instead of the left edge of the bucket.