do not close non-closable docks that are floating
Non-closable docks that are floating can still be closed by closing their window. This is a pretty serious bug as it may render the application completely unusable when the improperly closed docks contain important widgets.

This patch fixes the problem by moving non-closable docks to the home area when they are floating and their window is closed.