Make the documentation reproducible
Whilst working on the Reproducible Builds effort [0] we noticed that pyqtgraph
could not be built reproducibly. This is because it generates copyright years
from the current build date and therefore will vary on when you build it.

This commit uses SOURCE_DATE_EPOCH [1] for the "current" build date.

This was originally filed in Debian as #963124 [2].

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/
 [2] https://bugs.debian.org/963124