Move PyOpenGL import first and add # noqa
There are places throughout the library that the order of imports for
PyOpenGL and other OpenGL related modules is important. In all cases,
PyOpenGL should be imported first.  This change moves it to the top of
the file, and adds the # noqa comment so that isort does not rearrange
the ordering causing unintentional breakages.