Be lazier about importing h5py
We discovered recently bad interactions between pytables and h5py resulting in the inability to import one after the other had been imported. This takes a step towards pyqtgraph not being a confounding variable in that situation when it doesn't need to be. Note that h5py is still imported in MetaArray, so until that is removed, this doesn't actually solve the issue of h5py being loaded when pyqtgraph is imported, but once metaarray is removed as planned, this is the only remaing library (non-example) code that uses h5py directly.