avoid automatic conversion from float64 to float32
if we are going to use the float32 version of glVertexPointer, then we
should just pass in a float32 array. otherwise there will be an
automatic conversion by PyOpenGL.

on the other hand, if the intention of using float64 arrays was to have
float64 precision, then the use of glVertexPointf would be a bug.