Workaround for signal-method inheritance conflict
This fixes an issue that came up in PySide6 6.2.2 where calls to
GraphicsObject.parentChanged were passed to
QGraphicsObject.parentChanged instead of GraphicsItem.parentChanged. In
addition, this fix handles the case of ScaleBar.parentChanged as well,
where specifying the specific parent class was insufficient.

Thanks to @pijyoi for identifying this fix.