Remove Python 3.7 from CI - Add Conda Package Testing (#2211)
Remove Python 3.7 CI Pipelines

In order to conform to NEP-29, we now stop testing bindings on Py 3.7. 
Unfortunately, this means that we cannot test against PySide2 5.12 as the
pip installable version doesn't work with Python 3.8.  To test against
PySide2 5.12, we will have to add conda-based CI pipelines as the
conda-forge version of PySide2 has various fixes.

This PR also updates 

* `.pre-commit-config.yaml`
* `.gitignore`
* `pytest.ini` to filter out warnings specific to conda-forge/pyside2/python 3.8