Avoid all-NaNs warning
I'm uncertain whether the way is the best. If the data becomes all-NaNs, the bounds don't change, like if the plot is waiting for data.

With the fix, lines like `RuntimeWarning: All-NaN slice encountered ...` don't appear if the data consists of only NaNs.

And please make your mind up to whether `ScatterPlotItem.bounds` is a tuple or a list. I'd rather make it a tuple everywhere.