Allow passing in a pre-allocated path for reuse
If we don't reuse the `QPainterPath` that is be "reserved" (i.e. by
calling `.reserve()` on it to avoid re-allocating memory and using
`.clear()` instead of generating a new path) the mem allocs will be
somewhat pointless. Instead make it possible to pass in a previously
generated path, clear it, and rewrite to the reserved space.

Relevant docs:
- https://doc.qt.io/qt-5/qpainterpath.html#reserve
- https://doc.qt.io/qt-5/qpainterpath.html#clear