Interactive params fixup (#2318)
* Squashes all changes from #1812 into one commit to account for rebasing issues.

* apply black formatting

* Remove vestigial naming checks

* Bugfix: Parsing indented NumPy INI strings.
Uses smarter logic for determining how a docstring should be un-indented for INI parsing. Previously, param options nested under NumPy parameter specifications would be missed since they were under the indentation of a previous block. This new logic un-indents by checking for option indent level rather than relying on textwrap.dedent

* Greatly simplify `interact` logic.
Removes docstring parsing and `runFunc` behavior. The former is complex and error-prone, while the latter can be easily facilitated with a call to `functools.wraps` and doesn't need to pollute the interactive api.

* Fixup interactive formatting docs issues:
- Unicode instead of ASCII quotes/em-dashes
- 3-space instead of 4-space indents
- Case issues ("runOPts" vs "runOpts")
- Long one-line paragraphs rather than breaking by wrap boundary

* Improve `InteractiveFunction.__call__` logic.
Previously, when a `sigValue{Changed,Changing}` fired, this value would once again be propagated back to that Parameter before the function ran. This is just a minor performance overhead for `sigValueChanged`, since an equality check on the new value means it is a no-op. However, `sigValueChanging` shouldn't update the value of the parameter at all. Forcing `propagateChanges` to `False` prevents this.

* Much better interactive.py organization.
Utilize OOP to turn `interact` into an object method and greatly reduce the number of argument parameters for most helper methods. Has the nice side-benefit of better representing the state of defaults across different function calls, and lets users spawn their own `Interactor` classes if custom behavior is needed

* Minor cleanup: Fix codeql alerts + black formatting

* 100% `interactive` feature set test coverage

* Expand abbreviations: `param` -> `parameter`, etc.

* Fix incorrect indentation in rst doc

* Clean up docstrings & comments
- Wrap long lines
- Double-backtick literal expressions for proper sphinx rendering
- Fix typos

* `button` -> `action` for behavioral clarity

* Remove unused variable in `interact` test coverage

* Allow `extra` kwargs to update during `__call__`

* Implements `setDisconnected` similar to `setVisible` from graphics classes

* Bugfix: setting `__call__ = interact` exhibits incorrect behavior in subclasses.
If a subclass overrides `interact`, the subclass's method won't be called from `__call__` unless this fix is added.

* Bugfix: Old `interact` logic couldn't handle class methods

* force CI rerun without caching old results

* `InteractiveFunction.parameters` should be a dict instead of list
- The initial choice of a list came from the outdated `weakref` usage and is no longer relevant
- `extras` is also a dict, so this would represent consistent usage
- Makes it easier to identify which runtime kwargs match self's parameters