remove test for live-ness of w.plotItem.getMenu()
the test is actually bogus. the current implementation is such that it
has no C++ children, thus qObjectTree() will return only one element.
Yet if the menu did actually have any children, the test would fail on
PyQt bindings; because the Python wrappers of the children are still
alive (but the wrapped C++ objects are deleted)