FIX: ImageView nframes method is too hardcoded
This is fixing the behviour of ImageView for 3 dimentional images, where stacking or time is not in 0 index of array shape.
as nframes is being called by other methods which are enbled only if axes t is recognised, there is no need to check for if axes['t] is None (then frame should be 1);

alternatively, would not it be better to remove this method and use where needed `self.tVals.shape[0]` ? From performance point of view that would be better, but seeing the checking `if self.image is None:` probably there is the reason for cases where would be 0 frames. Thus I had added at the end `return 1` and so `nframes()` should cover all possible cases also then called externally.