Fixup
The previous commit broke the behavior for locked aspect ratio and
no limits on the view-range.

The previous and thereby desired behavior is the following:
Decreasing the size of a view keeps the view range constant, i.e.
does implicitly zoom-out.
If, afterwards, the size is increased again, the view range first
stays constant (i.e. implicit zoom-in) until the previous range and
zoom-level is reached again. Then, further increasing of the size
extends the view range (i.e. keeping zoom-level constant).