append to sys.path instead of push
After sending a push request for the previous change in this
patch branch, I noticed the following in the build log for Read the Docs:

> PyQtGraph requires one of PyQt5, PyQt6, PySide2 or PySide6; none of these packages could be imported.

Of course PyQt6 is listed in doc/requirements.txt and it was installed under that step:

> python -m pip install --exists-action=w --no-cache-dir -r doc/requirements.txt
> [...]
> Successfully installed PyQt6-6.6.0 [...]

Maybe this change produced some kind of a side effect for how the rtd
build resolves the imports for the docs build?

So here is a first guess, to try to unbreak it if the change
would break the docs build. Appending to sys.path instead of
adding the project root path to the front of sys.path