Allow release info change during apt-get update on ubuntu
The following was showing up in the build log, or some
CI builds under dependabot:

> Reading package lists...
> E: Repository 'https://packages.microsoft.com/ubuntu/22.04/prod jammy InRelease' changed its 'Origin' value from 'microsoft-ubuntu-jammy-prod jammy' to 'Pulp 3'
> E: Repository 'https://packages.microsoft.com/ubuntu/22.04/prod jammy InRelease' changed its 'Label' value from 'microsoft-ubuntu-jammy-prod jammy' to ''
> Error: Process completed with exit code 100.

apt-get(8) documentation[1] describes the --allow-releaseinfo-change
arg. This was also mentioned at Ask Ubuntu[2]

This patch updates the test-pip and test-conda jobs for the gh workflow,
adding the --allow-releaseinfo-change arg for the apt-get update
commands. This update is run before installing the Linux VirtualDisplay
dependencies.

[1]: https://manpages.debian.org/unstable/apt/apt-get.8.en.html
[2]: https://askubuntu.com/a/1093202