ViewBox: Only update auto-range before painting.
Instead of recalculating autorange for each call to addItem, defer the calculation until the next paint event.

I was experiencing a lot of lag when creating >1000 curve items with 100 points each. Profiling showed that 70% of the runtime was spent in updateAutoRange.
With this patch, updateAutoRange is less than 1% of the runtime.