Do not move PlotCurveItem to origin
The way I previously decided to fix #2661 was to shift PlotCurveItems
to be between (-1, 1).  There are other complications with shifting the
curve about the origin using PlotCurveItem.setData().  Furthermore, that
fix only addressed the case of PlotCurveItem, and not other
GraphicsItems.

An issue did come up with after doing an SVG export, where the
PlotCurveItem was moved out of the view with log-mode was enabled.

The correct fix is likely to apply a transform to the GraphicsItems so
they are positioned about the origin, and then have the QPainter
instance look there, instead of where the view is.  More experimentation
would be needed to make that work.