test_PolyLineROI now passes on non-AMD64 platforms
The major source of the discrepancy of the images came from AntiAliasing
This commit adds an attribute to ROIs and Handles, ._antialias, which is
True by default, but anti-aliasing will be disabled when that attribute
is set to False. The ROIs and Handles used in the test suite are set to
have this attribute set to False, thus disabling the use of
anti-aliasing.

This change also required that I generate new baseline test images,
which are included in the commit.

Because of disabling anti-aliasing, the pxCount value for image
comparison checker are substantially lower.