Use deploy script to install mesa3d drivers
Previous version of mesa3d setup involved registry hacks and copying
files. It did not work with anything newer than version 19.2.7. This
commit makes it so not only do we use the provided deploy script but we
are able to use the current version of mesa3d drivers as well.