Tidy up some tests, use more pytest.parametrize
Instead of using for-loops inside tests, use pytest.mark.parametrize
more.  In the past, this has been shown to have better resource
management and segfaults in the test suite have been avoided.  Also the
test output is easier to read on failure, and easier to recognize when
a supported combination of arguments is not being tested.