fix: num colors should equal num vertices
in the original code, we are generating faceCount colors (400) which i
more than the number of vertices (220).

the number of colors used by the OpenGL code is exactly the number of
vertices. (or in other words, each vertex has its own color)

the code still runs, just that only slightly more than half the generated
colors are used. it seems like this was not intended.