Activate propagation in containerChanged of Dock again
propagate was set to False to prevent the unintended collapse of a container tree with a single dock in a container.
But this lead to temporary areas closing anymore.
Instead, the removal of superfluous or empty containers is now only triggered by calling apoptose on the container if one of these cases is really present.
A unit test where this problem was discovered is added.